Refactor common class SerializableObject
The concept of serialization was inconsistent until now:
- To deserialize elements, the XML element tag names were defined
  outside the classes to be deserialized (e.g. the string "pin" was
  defined in the SymbolPin's outer class Symbol).
- To serialize elements, the classes to be serialized defined their XML
  element tag names in their own function serializeToXmlDomElement()
  (e.g. the string "pin" was defined in the class SymbolPin).

So nearly every XML tag name was defined in two different classes. To
make this more consistent, XML tag names are now always defined in the
outer classes (as described in the first point above) and are passed to
the inner classes as argument of serializeToXmlDomElement(). This way it
is also possible to use different XML tag names for the same class, for
example if a common/generic class is used in very different contexts.

Other changes:
- Make serializeToXmlDomElement() no longer pure virtual and implement
  it in SerializableObject instead. From now on this is only a
  convenience method.
- Add the pure virtual method serialize(XmlDomElement& root). In
  contrast to serializeToXmlDomElement() this new method allows to
  serialize multiple objects into the same DOM element (this is useful
  in some cases, e.g. to serialize a list of objects of mixed types).
- Add static helper template methods serializeObjectContainer() and
  serializePointerContainer() to make serialization of containers much
  more easy.
- Remove checkAttributesValidity() because it was bad design (see #157).